requirements.txt
ansitable==0.9.10
appdirs==1.4.4
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
-e git+ssh://git@github.com/lbusellato/CoLeCT.git@a5c307bba7455dc5ad407b051db178a01091485c#egg=CoLeCT
colored==1.4.4
contourpy==1.1.1
cycler==0.12.0
distlib==0.3.8
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.37
h5py==3.10.0
identify==2.5.33
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.0
lxml==5.1.0
matplotlib==3.7.3
netifaces==0.11.0
nodeenv==1.8.0
ntplib==0.4.0
numba==0.58.0
numpy==1.24.4
numpy-quaternion==2023.0.2
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
pgraph-python==0.6.2
Pillow==10.0.1
platformdirs==4.1.0
pre-commit==3.6.0
progress==1.6
protobuf==4.24.3
psutil==5.9.5
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytransform3d==3.5.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
roboticstoolbox-python==1.1.0
rtb-data==1.0.1
scikit-learn==1.3.1
scipy==1.10.1
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
spatialgeometry==1.1.0
spatialmath-python==1.1.9
swift-sim==1.1.0
threadpoolctl==3.2.0
tslearn==0.6.2
typing_extensions==4.8.0
tzdata==2023.3
ur-rtde==1.5.7
urllib3==2.0.6
virtualenv==20.25.0
wandb==0.15.11
websockets==12.0
zipp==3.17.0
